Перебрана система логирования, внесены пометки в код
